+++ index.html
@@ -4,6 +4,7 @@
 		<title>ClickPalette</title>
 		<link rel="stylesheet" type="text/css" href="./css/style.css">
 		<script>window.$ = window.jQuery = require('./js/jquery.min.js');</script>
+		<script src="./js/functions.js"></script>
 		<script src="./js/onload.js"></script>
 		<script src="./js/jqueryui.min.js"></script>
 	</head>
@@ -38,3 +39,4 @@
 
 	</body>
 </html>
+€
